REFAC: Extract the ingestion router to it's own module
